hdfs_ingestion/requirements.txt
kafka-python
hdfs
